+++ svc/FetchService.ts
@@ -15,6 +15,7 @@ import {
 } from '@xh/hoist/core';
 import {never, PromiseTimeoutSpec} from '@xh/hoist/promise';
 import {isLocalDate, olderThan, ONE_MINUTE, SECONDS} from '@xh/hoist/utils/datetime';
+import {apiDeprecated} from '@xh/hoist/utils/js';
 import {StatusCodes} from 'http-status-codes';
 import {isDate, isFunction, isNil, omitBy} from 'lodash';
 import {IStringifyOptions, stringify} from 'qs';
@@ -45,15 +46,25 @@ export class FetchService extends HoistService {
     NO_JSON_RESPONSES = [StatusCodes.NO_CONTENT, StatusCodes.RESET_CONTENT];
 
     private autoAborters = {};
-    defaultHeaders = {};
+    defaultHeaders: (PlainObject | ((arg: FetchOptions) => PlainObject))[] = [];
     defaultTimeout = (30 * SECONDS) as any;
 
     /**
      * Set default headers to be sent with all subsequent requests.
      * @param headers - to be sent with all fetch requests, or a function to generate.
+     * @deprecated
      */
     setDefaultHeaders(headers: PlainObject | ((arg: FetchOptions) => PlainObject)) {
-        this.defaultHeaders = headers;
+        apiDeprecated('setDefaultHeaders', {v: '66', msg: 'Use addDefaultHeaders instead'});
+        this.addDefaultHeaders(headers);
+    }
+
+    /**
+     * Add default headers to be sent with all subsequent requests.
+     * @param headers - to be sent with all fetch requests, or a function to generate.
+     */
+    addDefaultHeaders(headers: PlainObject | ((arg: FetchOptions) => PlainObject)) {
+        this.defaultHeaders.push(headers);
     }
 
     /**
@@ -147,16 +158,20 @@ export class FetchService extends HoistService {
     // Implementation
     //-----------------------
     private withDefaults(opts: FetchOptions, extraHeaders: PlainObject = null): FetchOptions {
-        const {defaultHeaders} = this,
-            method = opts.method ?? (opts.params ? 'POST' : 'GET'),
+        const method = opts.method ?? (opts.params ? 'POST' : 'GET'),
             isPost = method === 'POST';
+
+        const defaultHeaders = {};
+        this.defaultHeaders.forEach(h => {
+            Object.assign(defaultHeaders, isFunction(h) ? h(opts) : h);
+        });
 
         return {
             ...opts,
             method,
             headers: {
                 'Content-Type': isPost ? 'application/x-www-form-urlencoded' : 'text/plain',
-                ...(isFunction(defaultHeaders) ? defaultHeaders(opts) : defaultHeaders),
+                ...defaultHeaders,
                 ...extraHeaders,
                 ...opts.headers
             }